+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+- Add tests for caching functionality
+- Add option to disable caching
+- [Bugfix] template id caching was inconsistent
+
 ## 3.6.1 (June 15, 2020)
 
 - Code cleanup